feat: 提升模型精度为float32以改善转录准确度
将Whisper模型的计算精度从int8调整为float32，以提高转录的准确性。
这会增加内存使用量但会显著改善语音识别的精确度。

+++ faster_whisper_api.py
@@ -671,7 +671,7 @@ async def startup_event():
     print(f"Initializing {model_size} model on device {device}...")
     import time
     model_init_start = time.time()
-    model = WhisperModel(model_size, device=device, compute_type="int8")
+    model = WhisperModel(model_size, device=device, compute_type="float32")
     model_init_time = time.time() - model_init_start
     print(f"Model initialized successfully in {model_init_time:.2f}s!")
